simd: seperate non-rec context simd fron rec/simd.rs
Moved device-specific SIMD config code from rec/simd.rs
to the upper directory.
Now, rec/simd.rs only contains REC context-specific code.
